Schedule: Support additional video locations
Add an additional optional property called `otherUrls` that expects to
be an array of strings (URLs).

The code assumes Twitch, which is fine for simulstreaming at the moment.
It will need to be changed to properly support more URLs in the case of
collabs where the collab partners aren't on Twitch.